Fix qint64/int64_t template mismatches (long long vs long on Linux)
qMin/qMax/std::min/std::max need identical argument types; MSVC maps
both qint64 and int64_t to long long so this only surfaces on gcc/clang

+++ Src/TsFileDocument.cpp
@@ -943,8 +943,8 @@ public slots:
             }
             else
             {
-                firstTs_ = std::min(firstTs_, minTs);
-                lastTs_ = std::max(lastTs_, maxTs);
+                firstTs_ = std::min(firstTs_, qint64(minTs));
+                lastTs_ = std::max(lastTs_, qint64(maxTs));
             }
         }
